Estimate a relative number of cells per spot for spatial transcriptomics data

This function takes a list object containing spatial transcriptomics matrix as input and estimates the relative number of cells per spot in the data. The number of cells is estimated as the library size scaled to the range from 1 to 100. This value stored in the number_cells attribute.

Usage

get_num_cell_per_spot(alldata)

Arguments

alldata

A list object containing spatial transcriptomics

Value

a vector with the relative number of cells in each spot.

Examples

data("example_scrnaseq" , package = "scFeatures")
data <- example_scrnaseq
#> Loading required package: SeuratObject
#> Loading required package: sp
#> 
#> Attaching package: 'SeuratObject'
#> The following object is masked from 'package:base':
#> 
#>     intersect
data$celltype <- NULL

number_of_cells <- get_num_cell_per_spot(data)
#> Error in (function (cond) .Internal(C_tryCatchHelper(addr, 1L, cond)))(structure(list(message = "'data' not found in this Seurat object\n ",     trace = structure(list(call = list(tryCatch(withCallingHandlers({        NULL        saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpFLcHLZ/callr-fun-19964b627c11b3"),             list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv,             quote = TRUE), file = "/tmp/RtmpFLcHLZ/callr-res-19964b4b6463",             compress = FALSE)        flush(stdout())        flush(stderr())        NULL        invisible()    }, error = function(e) {        {            callr_data <- as.environment("tools:callr")$`__callr_data__`            err <- callr_data$err            if (FALSE) {                assign(".Traceback", .traceback(4), envir = callr_data)                dump.frames("__callr_dump__")                assign(".Last.dump", .GlobalEnv$`__callr_dump__`,                   envir = callr_data)                rm("__callr_dump__", envir = .GlobalEnv)            }            e <- err$process_call(e)            e2 <- err$new_error("error in callr subprocess")            class(e2) <- c("callr_remote_error", class(e2))            e2 <- err$add_trace_back(e2)            cut <- which(e2$trace$scope == "global")[1]            if (!is.na(cut)) {                e2$trace <- e2$trace[-(1:cut), ]            }            saveRDS(list("error", e2, e), file = paste0("/tmp/RtmpFLcHLZ/callr-res-19964b4b6463",                 ".error"))        }    }, interrupt = function(e) {        {            callr_data <- as.environment("tools:callr")$`__callr_data__`            err <- callr_data$err            if (FALSE) {                assign(".Traceback", .traceback(4), envir = callr_data)                dump.frames("__callr_dump__")                assign(".Last.dump", .GlobalEnv$`__callr_dump__`,                   envir = callr_data)                rm("__callr_dump__", envir = .GlobalEnv)            }            e <- err$process_call(e)            e2 <- err$new_error("error in callr subprocess")            class(e2) <- c("callr_remote_error", class(e2))            e2 <- err$add_trace_back(e2)            cut <- which(e2$trace$scope == "global")[1]            if (!is.na(cut)) {                e2$trace <- e2$trace[-(1:cut), ]            }            saveRDS(list("error", e2, e), file = paste0("/tmp/RtmpFLcHLZ/callr-res-19964b4b6463",                 ".error"))        }    }, callr_message = function(e) {        try(signalCondition(e))    }), error = function(e) {        NULL        if (FALSE) {            try(stop(e))        }        else {            invisible()        }    }, interrupt = function(e) {        NULL        if (FALSE) {            e        }        else {            invisible()        }    }), tryCatchList(expr, classes, parentenv, handlers), tryCatchOne(tryCatchList(expr,         names[-nh], parentenv, handlers[-nh]), names[nh], parentenv,         handlers[[nh]]), doTryCatch(return(expr), name, parentenv,         handler), tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),         tryCatchOne(expr, names, parentenv, handlers[[1L]]),         doTryCatch(return(expr), name, parentenv, handler), withCallingHandlers({            NULL            saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpFLcHLZ/callr-fun-19964b627c11b3"),                 list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv,                 quote = TRUE), file = "/tmp/RtmpFLcHLZ/callr-res-19964b4b6463",                 compress = FALSE)            flush(stdout())            flush(stderr())            NULL            invisible()        }, error = function(e) {            {                callr_data <- as.environment("tools:callr")$`__callr_data__`                err <- callr_data$err                if (FALSE) {                  assign(".Traceback", .traceback(4), envir = callr_data)                  dump.frames("__callr_dump__")                  assign(".Last.dump", .GlobalEnv$`__callr_dump__`,                     envir = callr_data)                  rm("__callr_dump__", envir = .GlobalEnv)                }                e <- err$process_call(e)                e2 <- err$new_error("error in callr subprocess")                class(e2) <- c("callr_remote_error", class(e2))                e2 <- err$add_trace_back(e2)                cut <- which(e2$trace$scope == "global")[1]                if (!is.na(cut)) {                  e2$trace <- e2$trace[-(1:cut), ]                }                saveRDS(list("error", e2, e), file = paste0("/tmp/RtmpFLcHLZ/callr-res-19964b4b6463",                   ".error"))            }        }, interrupt = function(e) {            {                callr_data <- as.environment("tools:callr")$`__callr_data__`                err <- callr_data$err                if (FALSE) {                  assign(".Traceback", .traceback(4), envir = callr_data)                  dump.frames("__callr_dump__")                  assign(".Last.dump", .GlobalEnv$`__callr_dump__`,                     envir = callr_data)                  rm("__callr_dump__", envir = .GlobalEnv)                }                e <- err$process_call(e)                e2 <- err$new_error("error in callr subprocess")                class(e2) <- c("callr_remote_error", class(e2))                e2 <- err$add_trace_back(e2)                cut <- which(e2$trace$scope == "global")[1]                if (!is.na(cut)) {                  e2$trace <- e2$trace[-(1:cut), ]                }                saveRDS(list("error", e2, e), file = paste0("/tmp/RtmpFLcHLZ/callr-res-19964b4b6463",                   ".error"))            }        }, callr_message = function(e) {            try(signalCondition(e))        }), saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpFLcHLZ/callr-fun-19964b627c11b3"),             list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv,             quote = TRUE), file = "/tmp/RtmpFLcHLZ/callr-res-19964b4b6463",             compress = FALSE), do.call(do.call, c(readRDS("/tmp/RtmpFLcHLZ/callr-fun-19964b627c11b3"),             list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv,             quote = TRUE), `<fn>`(base::quote(`<fn>`), base::quote(`<named list>`),             envir = base::quote(`<env>`), quote = base::quote(TRUE)),         `<fn>`(pkg = base::quote(`<pkgdown>`), examples = base::quote(TRUE),             run_dont_run = base::quote(FALSE), seed = base::quote(1014),             lazy = base::quote(FALSE), override = base::quote(`<list>`),             install = base::quote(FALSE), preview = base::quote(FALSE),             new_process = base::quote(FALSE), devel = base::quote(FALSE),             cli_colors = base::quote(256L), pkgdown_internet = base::quote(TRUE)),         pkgdown::build_site(...), build_site_local(pkg = pkg,             examples = examples, run_dont_run = run_dont_run,             seed = seed, lazy = lazy, override = override, preview = preview,             devel = devel), build_reference(pkg, lazy = lazy,             examples = examples, run_dont_run = run_dont_run,             seed = seed, override = override, preview = FALSE,             devel = devel), purrr::map(topics, build_reference_topic,             pkg = pkg, lazy = lazy, examples_env = examples_env,             run_dont_run = run_dont_run), map_("list", .x, .f,             ..., .progress = .progress), with_indexed_errors(i = i,             names = names, error_call = .purrr_error_call, call_with_cleanup(map_impl,                 environment(), .type, .progress, n, names, i)),         withCallingHandlers(expr, error = function(cnd) {            if (i == 0L) {            }            else {                message <- c(i = "In index: {i}.")                if (!is.null(names) && !is.na(names[[i]]) &&                   names[[i]] != "") {                  name <- names[[i]]                  message <- c(message, i = "With name: {name}.")                }                else {                  name <- NULL                }                cli::cli_abort(message, location = i, name = name,                   parent = cnd, call = error_call, class = "purrr_error_indexed")            }        }), call_with_cleanup(map_impl, environment(), .type,             .progress, n, names, i), .f(.x[[i]], ...), withCallingHandlers(data_reference_topic(topic,             pkg, examples_env = examples_env, run_dont_run = run_dont_run),             error = function(err) {                msg <- c(paste0("Failed to parse Rd in ", topic$file_in),                   i = err$message)                abort(msg, parent = err)            }), data_reference_topic(topic, pkg, examples_env = examples_env,             run_dont_run = run_dont_run), run_examples(tags$tag_examples[[1]],             env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),             topic = tools::file_path_sans_ext(topic$file_in),             run_dont_run = run_dont_run), highlight_examples(code,             topic, env = env), downlit::evaluate_and_highlight(code,             fig_save = fig_save_topic, env = child_env(env),             output_handler = evaluate::new_output_handler(value = pkgdown_print)),         evaluate::evaluate(code, child_env(env), new_device = TRUE,             output_handler = output_handler), evaluate_call(expr,             parsed$src[[i]], envir = envir, enclos = enclos,             debug = debug, last = i == length(out), use_try = stop_on_error !=                 2L, keep_warning = keep_warning, keep_message = keep_message,             log_echo = log_echo, log_warning = log_warning, output_handler = output_handler,             include_timing = include_timing), timing_fn(handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler))), handle(ev <- withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler)), try(f, silent = TRUE),         tryCatch(expr, error = function(e) {            call <- conditionCall(e)            if (!is.null(call)) {                if (identical(call[[1L]], quote(doTryCatch)))                   call <- sys.call(-4L)                dcall <- deparse(call, nlines = 1L)                prefix <- paste("Error in", dcall, ": ")                LONG <- 75L                sm <- strsplit(conditionMessage(e), "\n")[[1L]]                w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L],                   type = "w")                if (is.na(w))                   w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L],                     type = "b")                if (w > LONG)                   prefix <- paste0(prefix, "\n  ")            }            else prefix <- "Error : "            msg <- paste0(prefix, conditionMessage(e), "\n")            .Internal(seterrmessage(msg[1L]))            if (!silent && isTRUE(getOption("show.error.messages"))) {                cat(msg, file = outFile)                .Internal(printDeferredWarnings())            }            invisible(structure(msg, class = "try-error", condition = e))        }), tryCatchList(expr, classes, parentenv, handlers),         tryCatchOne(expr, names, parentenv, handlers[[1L]]),         doTryCatch(return(expr), name, parentenv, handler), withCallingHandlers(withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), warning = wHandler,             error = eHandler, message = mHandler), withVisible(eval_with_user_handlers(expr,             envir, enclos, user_handlers)), eval_with_user_handlers(expr,             envir, enclos, user_handlers), eval(expr, envir,             enclos), eval(expr, envir, enclos), get_num_cell_per_spot(data),         colSums(alldata$data), alldata$data, `$.Seurat`(alldata,             "data"), x[[i, drop = TRUE]], `[[.Seurat`(x, i, drop = TRUE),         tryCatch(expr = arg_match(arg = i, values = meta.cols,             multiple = TRUE), error = function(e) {            abort(message = paste(paste(sQuote(x = setdiff(x = i,                 y = meta.cols)), collapse = ", "), "not found in this Seurat object\n",                 e$body), call = rlang::caller_env(n = 4L))        }), tryCatchList(expr, classes, parentenv, handlers),         tryCatchOne(expr, names, parentenv, handlers[[1L]]),         value[[3L]](cond), abort(message = paste(paste(sQuote(x = setdiff(x = i,             y = meta.cols)), collapse = ", "), "not found in this Seurat object\n",             e$body), call = rlang::caller_env(n = 4L))), parent = c(0L,     1L, 2L, 3L, 2L, 5L, 6L, 0L, 0L, 0L, 0L, 0L, 12L, 13L, 14L,     15L, 16L, 17L, 18L, 17L, 17L, 21L, 21L, 23L, 24L, 25L, 26L,     27L, 28L, 28L, 30L, 31L, 32L, 33L, 34L, 28L, 28L, 28L, 38L,     39L, 40L, 41L, 41L, 41L, 44L, 44L, 46L, 47L, 48L, 49L, 50L    ), visible = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE,     FALSE, FALSE, FALSE, FALSE), namespace = c("base", "base",     "base", "base", "base", "base", "base", "base", "base", "base",     "base", NA, "pkgdown", "pkgdown", "pkgdown", "purrr", "purrr",     "purrr", "base", "purrr", "pkgdown", "base", "pkgdown", "pkgdown",     "pkgdown", "downlit", "evaluate", "evaluate", "evaluate",     "evaluate", "base", "base", "base", "base", "base", "base",     "base", "evaluate", "base", "base", "scFeatures", "methods",     NA, "SeuratObject", NA, "SeuratObject", "base", "base", "base",     NA, "rlang"), scope = c("::", "local", "local", "local",     "local", "local", "local", "::", "::", "::", "local", "global",     "::", ":::", "::", "::", ":::", ":::", "::", ":::", "local",     "::", ":::", ":::", ":::", "::", "::", ":::", "local", "local",     "::", "::", "local", "local", "local", "::", "::", ":::",     "::", "::", "::", "local", NA, ":::", NA, ":::", "::", "local",     "local", NA, "::"), error_frame = c(FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,     FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE,     FALSE, FALSE, FALSE)), row.names = c(NA, -51L), version = 2L, class = c("rlang_trace",     "rlib_trace", "tbl", "data.frame")), parent = NULL, rlang = list(        inherit = TRUE), call = x[[i, drop = TRUE]]), class = c("rlang_error", "error", "condition"))): error in evaluating the argument 'x' in selecting a method for function 'colSums': 'data' not found in this Seurat object
#>