data pre-processing

This function takes a Seurat object as input and does data cleaning and pre-processing. For example, it replaces the "+" and "-" signs in the celltype column with "plus" and "minus", respectively. It also removes patients that have less than 10 cells across all cell types. If the normalise argument is set to TRUE, the function will normalize the data using the Seurat::NormalizeData function.

process_data(data, normalise = TRUE)

Arguments

data

input data, a Seurat object.

normalise

a logical value indicating whether to normalize the data or not. Default value is TRUE.

Value

a Seurat object

Examples

data <- readRDS(system.file(
    "extdata", "example_spatial_transcriptomics.rds",
    package = "scFeatures"
))
data <- process_data(data, normalise = FALSE)