Create an association study report in HTML format

This function takes the feature matrix generated by scFeatures as input and creates an HTML report containing the results of the association study. The report is saved to the specified output folder.

Usage

run_association_study_report(scfeatures_result, output_folder)

Arguments

scfeatures_result

a named list storing the scFeatures feature output. Note that the names of the list should be one or multiple of the following: proportion_raw, proportion_logit, proportion_ratio, gene_mean_celltype, gene_prop_celltype, gene_cor_celltype, pathway_gsva, pathway_mean, pathway_prop, CCI, gene_mean_aggregated, gene_cor_aggregated, and gene_prop_aggregated.

output_folder

the path to the folder where the HTML report will be saved

Value

an HTML file, saved to the directory defined in the output_folder

argument

Examples

output_folder <- tempdir()
data("scfeatures_result" , package = "scFeatures")
run_association_study_report(scfeatures_result, output_folder )
#> 
#> 
#> processing file: output_report.Rmd
#> 1/89                   
#> 2/89 [unnamed-chunk-1] 
#> 3/89                   
#> 4/89 [unnamed-chunk-2] 
#> 5/89                   
#> 6/89 [unnamed-chunk-3] 
#> 7/89                   
#> 8/89 [unnamed-chunk-4] 
#> 9/89                   
#> 10/89 [unnamed-chunk-5] 
#> 11/89                   
#> 12/89 [unnamed-chunk-6] 
#> 13/89                   
#> 14/89 [unnamed-chunk-7] 
#> 15/89                   
#> 16/89 [unnamed-chunk-8] 
#> 17/89                   
#> 18/89 [unnamed-chunk-9] 
#> 19/89                   
#> 20/89 [unnamed-chunk-10]
#> 21/89                   
#> 22/89 [unnamed-chunk-11]
#> 23/89                   
#> 24/89 [unnamed-chunk-12]
#> 25/89                   
#> 26/89 [unnamed-chunk-13]
#> 27/89                   
#> 28/89 [unnamed-chunk-14]
#> 29/89                   
#> 30/89 [unnamed-chunk-15]
#> 31/89                   
#> 32/89 [unnamed-chunk-16]
#> 33/89                   
#> 34/89 [unnamed-chunk-17]
#> 35/89                   
#> 36/89 [unnamed-chunk-18]
#> 37/89                   
#> 38/89 [unnamed-chunk-19]
#> 39/89                   
#> 40/89 [unnamed-chunk-20]
#> 41/89                   
#> 42/89 [unnamed-chunk-21]
#> 43/89                   
#> 44/89 [unnamed-chunk-22]
#> 45/89                   
#> 46/89 [unnamed-chunk-23]
#> 47/89                   
#> 48/89 [unnamed-chunk-24]
#> 49/89                   
#> 50/89 [unnamed-chunk-25]
#> 51/89                   
#> 52/89 [unnamed-chunk-26]
#> 53/89                   
#> 54/89 [unnamed-chunk-27]
#> 55/89                   
#> 56/89 [unnamed-chunk-28]
#> 57/89                   
#> 58/89 [unnamed-chunk-29]
#> 59/89                   
#> 60/89 [unnamed-chunk-30]
#> 61/89                   
#> 62/89 [unnamed-chunk-31]
#> 63/89                   
#> 64/89 [unnamed-chunk-32]
#> 65/89                   
#> 66/89 [unnamed-chunk-33]
#> 67/89                   
#> 68/89 [unnamed-chunk-34]
#> 69/89                   
#> 70/89 [unnamed-chunk-35]
#> 71/89                   
#> 72/89 [unnamed-chunk-36]
#> 73/89                   
#> 74/89 [unnamed-chunk-37]
#> 75/89                   
#> 76/89 [unnamed-chunk-38]
#> 77/89                   
#> 78/89 [unnamed-chunk-39]
#> 79/89                   
#> 80/89 [unnamed-chunk-40]
#> 81/89                   
#> 82/89 [unnamed-chunk-41]
#> 83/89                   
#> 84/89 [unnamed-chunk-42]
#> 85/89                   
#> 86/89 [unnamed-chunk-43]
#> 87/89                   
#> 88/89 [unnamed-chunk-44]
#> 89/89                   
#> output file: output_report.knit.md
#> /usr/lib/rstudio-server/bin/quarto/bin/tools/pandoc +RTS -K512m -RTS output_report.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/Rtmp8lCZn5/output_report.html --lua-filter /dora/nobackup/yuec/R/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /dora/nobackup/yuec/R/rmarkdown/rmarkdown/lua/latex-div.lua --embed-resources --standalone --variable bs3=TRUE --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_collapsed=1 --variable toc_print=1 --template /dora/nobackup/yuec/R/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --number-sections --variable theme=bootstrap --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /tmp/Rtmp8lCZn5/rmarkdown-str2af90e7857bb40.html --variable code_folding=hide --variable code_menu=1 
#> 
#> Output created: /tmp/Rtmp8lCZn5/output_report.html