Generate overall aggregated gene correlation

This function computes the correlation of gene expression across samples. The user can specify the genes of interest, or let the function use the top variable genes by default. The function supports scRNA-seq, spatial proteomics, and spatial transcriptomics.

run_gene_cor(
  data,
  type = "scrna",
  genes = NULL,
  num_top_gene = NULL,
  ncores = 1
)

Arguments

data

A Seurat object containing celltype and sample label

type

The type of dataset, either "scrna", "spatial_t", or "spatial_p".

genes

Default to NULL, in which case the top variable genes will be used. If provided by user, need to be in the format of a list containing the genes of interest, eg, genes <- c(GZMA", "GZMK", "CCR7", "RPL38" )

num_top_gene

Number of top variable genes to use when genes is not provided. Defaults to 5.

ncores

Number of cores for parallel processing.

Value

a matrix of samples x features

Examples

if (FALSE) {
 data <- readRDS(
   system.file("extdata", "example_scrnaseq.rds", package = "scFeatures")
 )
 feature_gene_cor <- run_gene_cor(
   data, type = "scrna", num_top_gene = 5, ncores = 1
 )
}