wrapper function to run all feature types in scFeatures

wrapper function to run all feature types in scFeatures

scFeatures(
  data,
  feature_types = NULL,
  type = "scrna",
  ncores = 1,
  species = "Homo sapiens",
  celltype_genes = NULL,
  aggregated_genes = NULL,
  geneset = NULL,
  sample = "sample",
  celltype = "celltype",
  assay = "logcounts",
  spatialCoords = NULL
)

Arguments

data

input data, a Seurat object containing "sample" and "celltype" column. "x_cord" and "y_cord" is also required for constructing the features in the spatial metrics category.

feature_types

vector containing the name of the feature types to generate, options are "proportion_raw", "proportion_logit" , "proportion_ratio", "gene_mean_celltype", "gene_prop_celltype", "gene_cor_celltype", "pathway_gsva" , "pathway_mean", "pathway_prop", "gene_mean_aggregated", "gene_prop_aggregated", 'gene_cor_aggregated', "L_stats" , "celltype_interaction" , "morans_I", "nn_correlation". If no value is provided, all the above feature types will be generated.

type

input data type, either "scrna" (stands for single-cell RNA-sequencing data), "spatial_p" (stands for spatial proteomics data), or "spatial_t" (stands for single cell spatial data )

ncores

number of cores , default to 1

species

either "Homo sapiens" or "Mus musculus". Defaults to "Homo sapiens" if no value provided

celltype_genes

the genes of interest for celltype specific gene expression feature category If no value is provided, the top variable genes will be used

aggregated_genes

the genes of interest for overall aggregated gene expression feature category If no value is provided, the top variable genes will be used

geneset

the geneset of interest for celltype specific pathway feature category If no value is provided, the 50 hallmark pathways will be used

sample

the sample identifier if using a SingleCellExperiment

celltype

the celltype identifier if using a SingleCellExperiment

assay

the assay identifier if using a SingleCellExperiment

spatialCoords

the spatialCoords identifiers if using a SingleCellExperiment

Value

a list of dataframes containing the generated feature matrix in the form of sample x features

Examples

if (FALSE) {
 data <- readRDS(system.file("extdata", "example_scrnaseq.rds", package = "scFeatures"))
 scfeatures_result <- scFeatures(data, type = "scrna")
}